simulator/requirements.txt
paho-mqtt==1.6.1
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2